Adds EntityDocumentImpl and StatementDocumentImpl as utility classes
LexemeDocumentImpl is going to extend StatementDocumentImpl.
EntityDocumentImpl is here to have a top for the hierarchy that could
dispatch deserialization
